A better way to get word presence data from docs
The use of regular expressions for merely searching substrings
case-insensitively was an overkill. Moreover, this was not correct because of
not taking word delimiting into account (e. g., it would find “cat” in
“catalyst”).

The resulting matrix differs in one cell because of not discovering “interface”
in “interfaces” (which was just a happy coincidence).